Refactor: move the preview stuff to its own file
