+++ pyiron_workflow/function.py
@@ -5,7 +5,8 @@ from typing import Any, Literal, Optional, TYPE_CHECKING
 
 from pyiron_workflow.channels import InputData
 from pyiron_workflow.injection import OutputDataWithInjection
-from pyiron_workflow.io import Inputs, Outputs, ScrapesIO
+from pyiron_workflow.io import Inputs, Outputs
+from pyiron_workflow.io_preview import ScrapesIO
 from pyiron_workflow.node import Node
 from pyiron_workflow.snippets.colors import SeabornColors
