+++ pyiron_workflow/macro.py
@@ -13,7 +13,8 @@ import warnings
 from pyiron_workflow.channels import InputData, OutputData
 from pyiron_workflow.composite import Composite
 from pyiron_workflow.has_interface_mixins import HasChannel
-from pyiron_workflow.io import Outputs, Inputs, ScrapesIO
+from pyiron_workflow.io import Outputs, Inputs
+from pyiron_workflow.io_preview import ScrapesIO
 
 if TYPE_CHECKING:
     from pyiron_workflow.channels import Channel
